Revert "[C8-17] Improve portability of JVM source code when using gcc7."
This reverts commit ad818daa61ce15a9d066222cd0e4d83e3cc374c4.

+++ src/jdk/src/share/native/java/lang/fdlibm/src/s_floor.c
@@ -1,5 +1,6 @@
+
 /*
- * Copyright (c) 1998, 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1998, 2001, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -78,7 +79,7 @@ static double huge = 1.0e300;
                     if(j0==20) i0+=1;
                     else {
                         j = i1+(1<<(52-j0));
-                        if(j<(unsigned)i1) i0 +=1 ;       /* got a carry */
+                        if(j<i1) i0 +=1 ;       /* got a carry */
                         i1=j;
                     }
                 }
